# swagger-spec/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Register berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Data inputan tidak valid atau Email/Nomor Telepon sudah terdaftar"
          }
        },
        "summary": "Membuat akun pengguna baru (Public: Customer / Mitra)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil, mengembalikan token"
          },
          "401": {
            "description": "Email/Password salah atau akun ditangguhkan"
          }
        },
        "summary": "Masuk ke aplikasi",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Data pengguna ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token tidak valid atau belum dikirimkan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan data pengguna yang sedang login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Kata sandi berhasil diubah"
          },
          "401": {
            "description": "Kata sandi lama salah atau token tidak valid"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengubah kata sandi akun sendiri",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access Token baru berhasil diterbitkan"
          },
          "401": {
            "description": "Refresh token tidak valid atau kadaluwarsa"
          }
        },
        "summary": "Memperbarui Access Token menggunakan Refresh Token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout berhasil"
          },
          "401": {
            "description": "Token tidak valid"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Keluar dari aplikasi (Menghapus Refresh Token)",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/checkpoints/scan": {
      "post": {
        "operationId": "CheckpointsController_scanCheckpoint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCheckpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan checkpoint berhasil"
          },
          "400": {
            "description": "Pos tidak sesuai, OTP parcel salah, atau tiket tidak terdaftar di trip"
          },
          "404": {
            "description": "Data Trip atau Tiket tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Scan QR Checkpoint di Pos Asal/Tujuan (Operator Pos & Admin)",
        "tags": [
          "Checkpoints"
        ]
      }
    },
    "/api/checkpoints/manual-force-release": {
      "post": {
        "operationId": "CheckpointsController_manualForceRelease",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Pencairan manual berhasil"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Intervensi Darurat: Force Complete & Release Escrow manual oleh Operator Pos (HP Mitra Rusak)",
        "tags": [
          "Checkpoints"
        ]
      }
    },
...
    "/api/admin/dashboard/global": {
      "get": {
        "operationId": "AdminController_getGlobalDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Analitik global ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat dashboard analitik global (Superadmin Only)",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/dashboard/regional": {
      "get": {
        "operationId": "AdminController_getRegionalDashboard",
        "parameters": [
          {
            "name": "regionId",
            "required": false,
            "in": "query",
            "description": "Wajib diisi jika dikirim oleh Superadmin",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analitik wilayah ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat dashboard analitik wilayah",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/escrow/ledger": {
      "get": {
        "operationId": "AdminController_getEscrowLedger",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Buku besar Escrow ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat buku besar audit Escrow (Superadmin Only)",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/users/{id}/governance": {
      "patch": {
        "operationId": "AdminController_updateUserGovernance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserGovernanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status tata kelola user berhasil diperbarui"
          },
          "400": {
            "description": "Tidak dapat mengubah status sendiri"
          },
          "404": {
            "description": "User sasaran tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengubah status akun user (Suspended / Active / Banned)",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/settings/commission": {
      "patch": {
        "operationId": "AdminController_updatePlatformCommission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformCommissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persentase komisi platform berhasil diperbarui"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pengaturan persentase komisi platform global (Hanya Superadmin)",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/regions/{id}/rate": {
      "patch": {
        "operationId": "AdminController_updateRegionRate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegionRateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarif Rp/Km wilayah berhasil diperbarui"
          },
          "404": {
            "description": "Wilayah tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pengaturan tarif Rp/Km Wilayah Asal (Admin Wilayah & Superadmin)",
        "tags": [
          "Admin Governance"
        ]
      }
    },
    "/api/admin/settings/rewards": {
      "patch": {
        "operationId": "AdminController_updateRewardSetting",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pengaturan kelipatan Poin Reward berhasil diperbarui"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pengaturan kelipatan nominal Poin Reward global (Hanya Superadmin)",
        "tags": [
          "Admin Governance"
        ]
      }
    }
  },
  "info": {
    "title": "Nebeng App API Specification",
    "description": "Dokumentasi API Backend Platform Nebeng Transportasi & Pengiriman Paket (Escrow Wallet & Dual QR Checkpoint)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Masukkan token JWT: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1029103840"
          },
          "regionId": {
            "type": "object",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhon@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "1029391032211"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "regional",
              "operator",
              "mitra",
              "customer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ]
          },
          "statusVerification": {
            "type": "string",
            "enum": [
              "unverified",
              "pending",
              "approved",
              "rejected"
            ]
          },
          "avatar": {
            "type": "object",
            "example": "https://example.com/avatar.jpg"
          },
          "rewardPoints": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "role",
          "status",
          "statusVerification",
          "rewardPoints",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhons@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "09876543221"
          },
          "password": {
            "type": "string",
            "example": "Af^dfa1212daG"
          },
          "role": {
            "type": "string",
            "enum": [
              "customer",
              "mitra"
            ],
            "default": "customer"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "steve@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "P1ss^dhhc198"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "gshdjdhshfui"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "ScanCheckpointDto": {
        "type": "object",
        "properties": {
          "qrCodeTrip": {
            "type": "string",
            "example": "TRIP-A2D4CS13",
            "description": "Kode QR Trip milik mitra"
          },
          "qrCodeTicket": {
            "type": "string",
            "example": "TKT-SDJF12H",
            "description": "Kode QR Tiket / Parcel milik customer"
          },
          "posId": {
            "type": "string",
            "example": "10",
            "description": "ID pos resmi tempat operator bertugas"
          },
          "scanType": {
            "type": "string",
            "enum": [
              "checkin_origin",
              "checkin_destination"
            ],
            "example": "checkin_origin"
          },
          "otpClaim": {
            "type": "string",
            "example": "8765432",
            "description": "Kode OTP 6-digit milik penerima"
          },
          "securitySealQr": {
            "type": "string",
            "example": "SEAL-2981312",
            "description": "Nomor stiker segel QR"
          }
        },
        "required": [
          "qrCodeTrip",
          "qrCodeTicket",
          "posId",
          "scanType"
        ]
      },
...
      "UpdateUserGovernanceDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status akun baru (active, suspended, blocked)",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ],
            "example": "suspended"
          },
          "reason": {
            "type": "string",
            "description": "Alasan penanguhan atau pemblokiran akun",
            "example": "Pelanggaran ketetapan keamanan dan indikasi fraud"
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdatePlatformCommissionDto": {
        "type": "object",
        "properties": {
          "commissionPercentage": {
            "type": "number",
            "example": 10,
            "description": "presentase komisi platfrom"
          }
        },
        "required": [
          "commissionPercentage"
        ]
      },
      "UpdateRegionRateDto": {
        "type": "object",
        "properties": {
          "pricePerKm": {
            "type": "number",
            "example": 3000,
            "description": "Tarip dasar pengiriman/perjalanan per kilometer (Rp/Km)"
          }
        },
        "required": [
          "pricePerKm"
        ]
      }
    }
  }
}